Rename new chat app and legacy app files
